test e2e report refactor
