+++ src/modules/apps/e2e/index.tsx
@@ -2,7 +2,7 @@ import React from 'react'
 import { render } from 'ink'
 
 import { apps, tester } from '../../../clients'
-import { RealTimeReport } from './report/realTime'
+import { RealTimeReport } from './report/index'
 import { getToken } from '../../../conf'
 import { ManifestEditor } from '../../../lib/manifest/ManifestEditor'
 import { ErrorBoundary } from './ErrorBoundary'
